refactor: IconComponent 타입 공통 타입 파일로 이동
