+++ src/types/icon.ts
@@ -1,0 +1,3 @@
+import type { ComponentType, SVGProps } from 'react';
+
+export type IconComponent = ComponentType<SVGProps<SVGSVGElement>>;
